--- src/components/Header/ProjectHeader/ProjectVideoHeader/index.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ProjectVideoHeaderLayout from '../../../../layout/ProjectVideoHeader';
-
-function ProjectHeader() {
-  return (
-    <>
-      <ProjectVideoHeaderLayout />
-    </>
-  );
-}
-
-export default ProjectHeader;
