--- src/components/Header/ProjectHeader/ProjectVoiceHeader/index.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ProjectVoiceHeaderLayout from '../../../../layout/ProjectVoiceHeader';
-
-function ProjectHeader() {
-  return (
-    <>
-      <ProjectVoiceHeaderLayout />
-    </>
-  );
-}
-
-export default ProjectHeader;
